add default weather by ip

+++ domain/src/main/java/ru/ahoy/domain/models/Current.kt
@@ -70,5 +70,5 @@ data class Current(
     var gustMph: Double? = null,
 
     @SerializedName("gust_kph")
-    var gustKph: Int? = null
+    var gustKph: Double? = null
 )
